+++ src/Exception/MethodNotAllowedException.php
@@ -1,5 +1,10 @@
 <?php
 
+/**
+ * This file is part of d5whub extend router
+ * @author Vitor Reis <vitor@d5w.com.br>
+ */
+
 namespace D5WHUB\Extend\Router\Exception;
 
 class MethodNotAllowedException extends RuntimeException