+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG
 ---
 
 ## 4.1.1
+- Fix variable params
 - Performance improvements
   - Fix unique added route index in route collection variable tree
   - Fix match indexes tree